+++ src/pollard_p1_factorization.h
@@ -1,8 +1,8 @@
 #ifndef POLLARD_P1_FACTORIZATION_H
 #define POLLARD_P1_FACTORIZATION_H
 
-void pollard_p1_factorization(long long int n, long long int* p, long long int* q, unsigned int *primes, unsigned int primes_length);
+void pollard_p1_factorization(long long int n, long long int* p, long long int* q, unsigned long int *primes, unsigned long int primes_length);
 
-long long int pollard_p1_factor(long long int n, unsigned int *primes, unsigned int primes_length);
+long long int pollard_p1_factor(long long int n, unsigned long int *primes, unsigned long int primes_length);
 
 #endif
